backend/requirements.txt
django==5.1.3
psycopg2-binary
django-cors-headers
django-environ
djangorestframework
django-filter
Pillow
gunicorn